refactor(handler): relocate io_native and io_stub to core/context/io
