+++ packages/logd/lib/src/handler/encoder/encoder.dart
@@ -6,12 +6,13 @@ import 'dart:typed_data';
 import 'package:meta/meta.dart';
 
 import '../../../logd.dart' show ConsoleSink, LogSink;
+import '../../core/context/io/io_stub.dart'
+    if (dart.library.io) '../../core/context/io/io_native.dart' as io;
 import '../../core/theme/log_theme.dart';
 import '../../logger/logger.dart';
 import '../document/document.dart';
 import '../engine/engine.dart';
 import '../handler.dart' show ConsoleSink, LogSink;
-import '../io_stub.dart' if (dart.library.io) '../io_native.dart' as io;
 import '../layout/layout.dart';
 import '../sink/sink.dart' show ConsoleSink, LogSink;
